requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
dataclasses-json==0.5.8
Django==4.2.2
django-environ==0.10.0
djangorestframework==3.14.0
faiss-cpu==1.7.4
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.208
langchainplus-sdk==0.0.16
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
psycopg2==2.9.6
pydantic==1.10.9
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
